Add loglevel option (see commit description)
As mentioned in [1], when prettier formats at least one file, it always prints the paths of
all changed files, regardless whether it changed them. They come in different shades of grey,
but in a light theme they are almost undistinguishable. Therefore, we have to decide if we want
to see the paths of all files, or none and just a hint of pre-commit that files were changed by
this hook. I will keep it this way for now, as `git diff` will also list the changes.

[1]: https://github.com/prettier/prettier/issues/3525