Move collectstatic to entry point
This was necessary because the collection of statics depends on the
BACKEND configuration. We only have that in production config right
now. A (better?) alternative would be to add whitenoise in dev
configuration and collect statics during the build process.
Need to test if this will cause problems in dev as runserver
is used to serve the files and I need to check if that honors
the configured backend